........
  r1122 | jaroslavl1 | 2009-01-10 12:58:49 +0100 (so, 10 I 2009) | 1 line
  
  Added G.726 encoder/decoder wrapper in sipxmedialib. G.726 is now functional.
........
  r1123 | jaroslavl1 | 2009-01-10 13:00:02 +0100 (so, 10 I 2009) | 1 line
  
  Cleanup of IPP codecs, removed unused methods.
........
  r1124 | jaroslavl1 | 2009-01-10 14:50:41 +0100 (so, 10 I 2009) | 1 line
  
  Minor refactoring of DTMF detector code, to support additional DTMF detectors (new wikipedia, SpanDSP).
........
  r1125 | jaroslavl1 | 2009-01-10 16:08:17 +0100 (so, 10 I 2009) | 6 lines
  
  Merged DTMF detector from SIP foundry from Keith Kyzivat (based on new wikipedia algorithm) with old DTMF detector. This seems to provide best of the both worlds. Subjectively this DTMF detector performs better than the old DTMF detector (less false detections, and more correct detections).
  
  Tested with wxCommunicator (with sipxtapi calllib rewrite), SJPhone 1.65 and X-Lite. All inband DTMF tones were detected over G.711a codec from all softphones, with no false tones being detected. Old DTMF detector was unable to detect digit 4 and 5 from SJPhone, but new DTMF detector works correctly.
  It also seems to have higher resistance to echo.
  
  Modified license to add all involved parties.
........
  r1126 | jaroslavl1 | 2009-01-10 16:10:59 +0100 (so, 10 I 2009) | 1 line
  
  Removed old MicThreadWnt which is not used anymore.
........
  r1127 | jaroslavl1 | 2009-01-10 16:21:25 +0100 (so, 10 I 2009) | 1 line
  
  Parameter naming fix.
........
  r1128 | jaroslavl1 | 2009-01-10 17:29:24 +0100 (so, 10 I 2009) | 1 line
  
  Updated Span DSP projects to compile also DTMF detector. Fixed floating_fudge.h to work with VS2008. It seems math.h defines also functions which floating_fudge.h is trying to define.
........
  r1129 | jaroslavl1 | 2009-01-10 17:33:11 +0100 (so, 10 I 2009) | 4 lines
  
  Added MprSpanDspDtmfDetector, and enabled it by default. It can be disabled by removing USE_SPAN_DSP_DTMF definition from sipxmedialib projects properties.
  
  This detector seems to be performing the same or better as internal simple DTMF detector based on wikipedia algorithm.
........
  r1130 | jaroslavl1 | 2009-01-10 20:18:57 +0100 (so, 10 I 2009) | 1 line
  
  Reduced size of MpJitterBufferDefault minimum prefetch count from 8 to 6, since 6 is the initial value. 6 is able to cover more than 100ms latency for 20ms frames. Reduced maximum value from 30 to 20, as 20 should be sufficient.
........
  r1133 | jaroslavl1 | 2009-01-11 00:55:49 +0100 (ne, 11 I 2009) | 3 lines
  
  Don't send CANCEL if we attempt to drop call in remote offering state, and INVITE transaction is not active for some reason (bug..), just drop the call.
  
  When force destroying a connection, use CP_CALLSTATE_CAUSE_NO_RESPONSE as cause.
........
  r1135 | jaroslavl1 | 2009-01-11 13:08:17 +0100 (ne, 11 I 2009) | 1 line
  
  Removed SpeakerThreadWnt which is no longer used.
........
  r1136 | jaroslavl1 | 2009-01-11 13:14:42 +0100 (ne, 11 I 2009) | 1 line
  
  64bit fixes for RTCPTimer.
........
  r1138 | jaroslavl1 | 2009-01-11 17:31:40 +0100 (ne, 11 I 2009) | 1 line
  
  Improved handling of remote capability discovery (sending OPTIONS). Do not send OPTIONS if we are disconnecting.
........

+++ sipXmediaLib/src/rtcp/RTCPTimer.cpp
@@ -356,7 +356,7 @@ unsigned int __stdcall CRTCPTimer::TimerThreadProc(void * lpParameter)
  *
  *
  */
-void  CRTCPTimer::ReportingAlarm(timer_t tTimer, int iArgument)
+void  CRTCPTimer::ReportingAlarm(timer_t tTimer, intptr_t iArgument)
 {
     CRTCPTimer   *poRTCPTimer = (CRTCPTimer  *)iArgument;
 
@@ -370,7 +370,7 @@ void  CRTCPTimer::ReportingAlarm(timer_t tTimer, int iArgument)
 #include <sys/time.h>
 #endif
 
-void CRTCPTimer::ReportingAlarm(const int userData, const int eventData)
+void CRTCPTimer::ReportingAlarm(const intptr_t userData, const intptr_t eventData)
 {
 #ifdef RTCP_LINUX_DEBUG
     struct timeval tv;
